+++ flask_restful/inputs.py
@@ -1,6 +1,6 @@
 from calendar import timegm
 from datetime import datetime, time, timedelta
-from email.utils import formatdate, parsedate_tz, mktime_tz
+from email.utils import parsedate_tz, mktime_tz
 import re
 
 import aniso8601
@@ -215,36 +215,6 @@ def boolean(value):
     if value in ('false', '0',):
         return False
     raise ValueError("Invalid literal for boolean(): {}".format(value))
-
-
-def rfc822(dt):
-    """Turn a datetime object into a formatted date.
-
-    Example::
-
-        inputs.rfc822(datetime(2011, 1, 1)) => "Sat, 01 Jan 2011 00:00:00 -0000"
-
-    :param dt: The datetime to transform
-    :type dt: datetime
-    :return: A RFC 822 formatted date string
-    """
-    return formatdate(timegm(dt.utctimetuple()))
-
-
-def iso8601(dt):
-    """Turn a datetime object into an ISO8601 formatted date.
-
-    Example::
-
-        inputs.iso8601(datetime(2012, 1, 1, 0, 0)) => "2012-01-01T00:00:00+00:00"
-
-    :param dt: The datetime to transform
-    :type dt: datetime
-    :return: A ISO 8601 formatted date string
-    """
-    return datetime.isoformat(
-        datetime.fromtimestamp(timegm(dt.utctimetuple()), tz=pytz.UTC)
-    )
 
 
 def datetime_from_rfc822(datetime_str):
